# Use net.LookupTXT instead of dig for AppEngine ips