This pull request replaces the use of dig and `exec.Command` with `net.LookupTXT` in `DownloadAppEngine`. This is substantially faster and uses much less memory.

It's also far more reliable as it does not require an external binary and does not parse the output of a command. I also implemented a more rigorous parser of the spf record.

This pull request also adds a test case that simply compares the result against the old implementation, that now exists solely for tests.

Performance numbers are as follows:
```
name                 old time/op    new time/op    delta
DownloadAppEngine-8     2.14s ±19%     1.01s ±50%  -52.69%  (p=0.000 n=18+18)

name                 old alloc/op   new alloc/op   delta
DownloadAppEngine-8     363kB ± 1%      35kB ± 0%  -90.34%  (p=0.000 n=20+18)

name                 old allocs/op  new allocs/op  delta
DownloadAppEngine-8     1.17k ± 0%     0.40k ± 0%  -66.00%  (p=0.000 n=20+17)
```