@tmthrgd 2018-01-06
The travis build <https://travis-ci.org/client9/ipcat/builds/325777963> is failing because the dig command is not available. A possible solution is just to skip this test if dig isn't available. That wouldn't be any worse than before because `DownloadAppEngine` previously had no test case at all.