@shawnps 2018-01-17
@tmthrgd thanks for the PR. Let's just remove the references to the `dig` calls and use the `net. LookupTXT`. That way we don't need the tests and benchmarks around the old code.