@tmthrgd 2018-01-18
@shawnps So just to clarify, you want me to remove all of `appengine_test.go`?