@shawnps 2018-01-18
@tmthrgd I might be missing something, but it looks like those tests assume we're using `dig`, but the new functionality that you wrote would be using `net.LookupTXT`, so the tests wouldn't be testing the actual functionality anymore, right? If that's the case I say just remove the tests, yeah. What do you think?