@tmthrgd 2018-01-18
@shawnps `TestDownloadAppEngine` is a regression test, it compares the output of the new and old methods to make sure the results are the same.

It's not strictly needed, and it's very much up to you whether I remove it or not.