@shawnps 2018-01-19
@tmthrgd that's okay, using those commands to update the appengine data is a manual process so if anything goes wrong I'll probably notice. Feel free to remove the regression test